Database.Drasil.ChunkDB

Documentation

basisCDB :: ChunkDB

The basis chunk database, which contains the basic idea dicts, concept chunks, and units that are used in all of the case studies. This database is then added to all of the new chunk databases created using the cdb constructor.