Drasil.Generator

Base ChunkDB for all case studies

cdb :: [DefinedQuantityDict] -> [IdeaDict] -> [ConceptChunk] -> [UnitDefn] -> [DataDefinition] -> [InstanceModel] -> [GenDefn] -> [TheoryModel] -> [ConceptInstance] -> [LabelledContent] -> [Reference] -> [Citation] -> ChunkDB

Tools for dumping a chunk database to disk for debugging

dumpEverything :: System -> PrintingInformation -> Path -> IO ()

For debugging purposes, if the system has a DEBUG_ENV environment variable set to anything, we can dump the chunk maps in a system to the host system.

Generator Functions

gen :: DocSpec -> Document -> PrintingInformation -> IO ()

Generate a number of artifacts based on a list of recipes.

genDot :: System -> IO ()

Generates traceability graphs as .dot files.

genCode :: Choices -> CodeSpec -> IO ()

Calls the code generator.

Types (Printing Options)

data DocType

Document types include Software Requirements Specification and Website. Choosing SRS will generate both TeX and HTML files, while Website generates only as HTML. This also determines what folders the generated files will be placed into.

Constructors

SRS

Website

Lesson

data DocSpec

Document specifications. Holds the type of document (DocType) and its name (Filename).

Constructors

DocSpec DocChoices Filename

data DocChoices

Document choices include the type of document as well as the file formats we want to generate as.

Constructors

DC DocType [Format]

data Format

Possible formats for printer output.

Constructors

TeX

HTML

Jupyter

MDBook

Constructor

docChoices :: DocType -> [Format] -> DocChoices

Constructor for users to choose their document options

Type check a Drasil System

typeCheckSI :: System -> IO ()