Language.Drasil.Dump

Documentation

type Path = String

type TargetFile = String

dumpEverything :: System -> PrintingInformation -> Path -> IO ()

For debugging purposes, if the system has a DEBUG_ENV environment variable set to anything, we can dump the chunk maps in a system to the host system.

dumpEverything0 :: System -> PrintingInformation -> Path -> IO ()

dumpTo :: ToJSON a => a -> TargetFile -> IO ()

dumpChunkTables :: PrintingInformation -> TargetFile -> IO ()