Language.Drasil.Output.Formats

Description

Defines output formats for the different documents we can generate.

Documentation

type Filename = String

When choosing your document, you must specify the filename for the generated output (specified without a file extension).

data DocChoices

Document choices include the type of document as well as the file formats we want to generate as.

Constructors

DC

Fields

doctype :: DocType

format :: [Format]

data DocSpec

Document specifications. Holds the type of document (DocType) and its name (Filename).

Constructors

DocSpec DocChoices Filename

data DocClass

LaTeX helper.

Constructors

DocClass (Maybe String) String

newtype UsePackages

LaTeX helper for adding packages. Wraps a list of package names.

Constructors

UsePackages [String]

data ExDoc

LaTeX helper.

Constructors

ExDoc (Maybe String) String